requirements.txt
future
emoji
beautifulsoup4
requests
sqlalchemy
python-telegram-bot==11.1.0
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
googletrans
jikanpy
speedtest-cli
coffeehouse
tswift
gTTS
wikipedia
pyowm
tabulate
parsel
telethon
pyrogram
git+https://github.com/ALENTL/bing_image_downloader
pySmartDL
hurry.filesize
fonttools